requirements.txt
# Empty
google-cloud-secret-manager
requests